Create useModalScreenOptions hook
